:recycle: [FLAV-78] Refactor BeanList
