+++ src/components/App.js
@@ -3,7 +3,6 @@ import { BrowserRouter, Route, Switch } from "react-router-dom";
 import AppBar from "./AppBar.js";
 import AuthFireRoute from "./AuthFireRoute.js";
 import { AuthProvider } from "./contexts/AuthContext.js";
-import BeansList from "../features/myBeans/BeansList.js";
 import Expert from "./Register/Expert.js";
 import ForgetPassword from "./ForgetPassword.js";
 import { GlobalStyles } from "../GlobalStyles";
@@ -11,6 +10,7 @@ import Home from "./Home.js";
 import Login from "./Login.js";
 import Member from "./Member.js";
 import MemberTest from "./MemberTest.js";
+import { Page as MyBeans } from "../features/myBeans/Page.js";
 import PropTypes from "prop-types";
 import React from "react";
 import Signup from "./Signup.js";
@@ -73,7 +73,7 @@ export default function App() {
                   <MemberTest />
                 </Route>
                 <Route path="/member-test2">
-                  <BeansList />
+                  <MyBeans />
                 </Route>
                 <Route path="/selection">
                   <Expert />
