+++ src/features/myBeans/beansSlice.js
@@ -1,6 +1,7 @@
 import {
   createAsyncThunk,
   createEntityAdapter,
+  createSelector,
   createSlice,
 } from "@reduxjs/toolkit";
 
@@ -39,3 +40,7 @@ export const {
   selectAll: selectBeans,
   selectById: selectBeanById,
 } = beansAdapter.getSelectors((state) => state.beans);
+
+export const selectBeanIds = createSelector(selectBeans, (beans) =>
+  beans.map((bean) => bean.id)
+);
